Move out-of-place files related to SQL.
