+++ metricflow/inference/renderer/stream.py
@@ -4,9 +4,9 @@ import sys
 from collections import defaultdict
 from typing import Dict, List, TextIO
 
-from metricflow.dataflow.sql_table import SqlTable
 from metricflow.inference.models import InferenceResult
 from metricflow.inference.renderer.base import InferenceRenderer
+from metricflow.sql.sql_table import SqlTable
 
 
 class StreamInferenceRenderer(InferenceRenderer):
