+++ metricflow/inference/solver/base.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 from abc import ABC, abstractmethod
 from typing import List
 
-from metricflow.dataflow.sql_column import SqlColumn
 from metricflow.inference.models import InferenceResult, InferenceSignal
+from metricflow.sql.sql_column import SqlColumn
 
 
 class InferenceSolver(ABC):
