+++ metricflow/model/semantic_manifest_lookup.py
@@ -5,12 +5,12 @@ import logging
 from dbt_semantic_interfaces.protocols.semantic_manifest import SemanticManifest
 from dbt_semantic_interfaces.type_enums import TimeGranularity
 
-from metricflow.dataflow.sql_table import SqlTable
 from metricflow.mf_logging.pretty_print import mf_pformat
 from metricflow.model.semantics.metric_lookup import MetricLookup
 from metricflow.model.semantics.semantic_model_lookup import SemanticModelLookup
 from metricflow.plan_conversion.time_spine import TimeSpineSource
 from metricflow.protocols.semantics import SemanticModelAccessor
+from metricflow.sql.sql_table import SqlTable
 
 logger = logging.getLogger(__name__)
