+++ metricflow/test/api/test_metricflow_client.py
@@ -3,9 +3,9 @@ from __future__ import annotations
 from dbt_semantic_interfaces.validations.validator_helpers import SemanticManifestValidationResults
 
 from metricflow.api.metricflow_client import MetricFlowClient
-from metricflow.dataflow.sql_table import SqlTable
 from metricflow.engine.models import Dimension, Metric
 from metricflow.random_id import random_id
+from metricflow.sql.sql_table import SqlTable
 from metricflow.test.fixtures.setup_fixtures import MetricFlowTestSessionState
