+++ metricflow/test/dataflow/test_sql_column.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
-from metricflow.dataflow.sql_column import SqlColumn
-from metricflow.dataflow.sql_table import SqlTable
+from metricflow.sql.sql_column import SqlColumn
+from metricflow.sql.sql_table import SqlTable
 
 
 def test_sql_column() -> None:  # noqa: D
