+++ metricflow/test/fixtures/sql_clients/base_sql_client_implementation.py
@@ -7,7 +7,6 @@ from typing import Optional
 
 import pandas as pd
 
-from metricflow.dataflow.sql_table import SqlTable
 from metricflow.mf_logging.formatting import indent
 from metricflow.mf_logging.pretty_print import mf_pformat
 from metricflow.protocols.sql_client import (
@@ -15,6 +14,7 @@ from metricflow.protocols.sql_client import (
 )
 from metricflow.random_id import random_id
 from metricflow.sql.sql_bind_parameters import SqlBindParameters
+from metricflow.sql.sql_table import SqlTable
 from metricflow.sql_request.sql_request_attributes import SqlRequestId
 
 logger = logging.getLogger(__name__)
