+++ metricflow/test/inference/context/test_snowflake.py
@@ -5,8 +5,6 @@ from unittest.mock import MagicMock
 
 import pandas as pd
 
-from metricflow.dataflow.sql_column import SqlColumn
-from metricflow.dataflow.sql_table import SqlTable
 from metricflow.inference.context.data_warehouse import (
     ColumnProperties,
     DataWarehouseInferenceContext,
@@ -14,6 +12,8 @@ from metricflow.inference.context.data_warehouse import (
     TableProperties,
 )
 from metricflow.inference.context.snowflake import SnowflakeInferenceContextProvider
+from metricflow.sql.sql_column import SqlColumn
+from metricflow.sql.sql_table import SqlTable
 
 
 def test_column_type_conversion() -> None:  # noqa: D
