+++ metricflow/test/inference/renderer/test_config_file.py
@@ -7,9 +7,9 @@ from pathlib import Path
 import pytest
 from ruamel.yaml import YAML
 
-from metricflow.dataflow.sql_column import SqlColumn
 from metricflow.inference.models import InferenceResult, InferenceSignalType
 from metricflow.inference.renderer.config_file import ConfigFileRenderer
+from metricflow.sql.sql_column import SqlColumn
 
 yaml = YAML()
