+++ metricflow/test/inference/rule/test_rules.py
@@ -1,7 +1,5 @@
 from __future__ import annotations
 
-from metricflow.dataflow.sql_column import SqlColumn
-from metricflow.dataflow.sql_table import SqlTable
 from metricflow.inference.context.data_warehouse import (
     ColumnProperties,
     DataWarehouseInferenceContext,
@@ -10,6 +8,8 @@ from metricflow.inference.context.data_warehouse import (
 )
 from metricflow.inference.models import InferenceSignalConfidence, InferenceSignalType
 from metricflow.inference.rule.rules import ColumnMatcherRule, LowCardinalityRatioRule
+from metricflow.sql.sql_column import SqlColumn
+from metricflow.sql.sql_table import SqlTable
 
 
 def create_context_with_counts(rows: int, distinct: int, nulls: int) -> DataWarehouseInferenceContext:
