+++ metricflow/test/inference/solver/test_weighted_tree.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 
-from metricflow.dataflow.sql_column import SqlColumn
 from metricflow.inference.models import InferenceSignal, InferenceSignalConfidence, InferenceSignalType
 from metricflow.inference.solver.weighted_tree import WeightedTypeTreeInferenceSolver
+from metricflow.sql.sql_column import SqlColumn
 
 column = SqlColumn.from_string("db.schema.table.col")
 solver = WeightedTypeTreeInferenceSolver()
