+++ metricflow/test/mf_logging/test_dag_to_text.py
@@ -8,12 +8,12 @@ from typing import List
 
 from metricflow.dag.dag_to_text import MetricFlowDagTextFormatter
 from metricflow.dag.mf_dag import DagId
-from metricflow.dataflow.sql_table import SqlTable
 from metricflow.mf_logging.formatting import indent
 from metricflow.sql.sql_exprs import (
     SqlStringExpression,
 )
 from metricflow.sql.sql_plan import SqlQueryPlan, SqlSelectColumn, SqlSelectStatementNode, SqlTableFromClauseNode
+from metricflow.sql.sql_table import SqlTable
 
 logger = logging.getLogger(__name__)
