+++ metricflow/test/table_snapshot/table_snapshots.py
@@ -15,8 +15,8 @@ import yaml
 from dbt_semantic_interfaces.enum_extension import assert_values_exhausted
 from dbt_semantic_interfaces.implementations.base import FrozenBaseModel
 
-from metricflow.dataflow.sql_table import SqlTable
 from metricflow.specs.specs import hash_items
+from metricflow.sql.sql_table import SqlTable
 from metricflow.test.fixtures.sql_clients.ddl_sql_client import SqlClientWithDDLMethods
 
 logger = logging.getLogger(__name__)
